feat(statistic): add StatisticCard component and related stories
